Moved FrameBlock and Barrier into thier own respective headers
