+++ include/vsg/viewer/Viewer.h
@@ -15,10 +15,9 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #include <vsg/viewer/Presentation.h>
 #include <vsg/viewer/RecordAndSubmitTask.h>
 #include <vsg/viewer/Window.h>
-
 #include <vsg/traversals/CompileTraversal.h>
-#include <vsg/ui/ApplicationEvent.h>
-#include <vsg/vk/Context.h>
+#include <vsg/threading/FrameBlock.h>
+#include <vsg/threading/Barrier.h>
 
 #include <map>
